refactor: move hardcoded servers configuration to config file
The gRPC and REST servers had hardcoded interface and port bindings.
In order to use docker, the interface needs to be different than the one
used when developing with local servers.
Moving the servers interface to a configuration file makes it easier to
override it via CLI parameters when executing from docker